Add report building (#216)
* Add report building

* Cleanup of PR comments, typing fix for KVStore to explicitly specify uint64 or float64 for precision purposes

* Fix zip() usage, remove overly strict guard for TPS value

* Move report_builder.py to be a classmethod in Report. Use 'lower' instead of linear interpolation for numpy percentile

+++ src/inference_endpoint/async_utils/services/metrics_aggregator/aggregator.py
@@ -114,7 +114,7 @@ class MetricsAggregatorService(ZmqEventRecordSubscriber):
         self._tracked_issued = 0
         self._tracked_completed = 0
         self._session_start_ns: int | None = None
-        self._total_duration_ns: float = 0.0
+        self._total_duration_ns: int = 0
 
         self._table = MetricsTable(kv_store)
         self._register_triggers(streaming)
